+++ mods/sdk/CHANGELOG.md
@@ -3,6 +3,20 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.19.1](https://github.com/fonoster/qcobro/compare/v1.19.0...v1.19.1) (2026-07-13)
+
+**Note:** Version bump only for package @qcobro/sdk
+
+# [1.19.0](https://github.com/fonoster/qcobro/compare/v1.18.0...v1.19.0) (2026-07-13)
+
+**Note:** Version bump only for package @qcobro/sdk
+
+# [1.18.0](https://github.com/fonoster/qcobro/compare/v1.17.0...v1.18.0) (2026-07-13)
+
+### Features
+
+- **mcp:** add MCP server package for the QCobro API ([#38](https://github.com/fonoster/qcobro/issues/38)) ([baeeb10](https://github.com/fonoster/qcobro/commit/baeeb1029df4d02b49f5b4204500d0be8306391a)), closes [fonoster/qcobro#16](https://github.com/fonoster/qcobro/issues/16)
+
 # [1.16.0](https://github.com/fonoster/qcobro/compare/v1.15.0...v1.16.0) (2026-07-13)
 
 **Note:** Version bump only for package @qcobro/sdk